+++ src/utils/chainId.utils.js
@@ -1,0 +1,11 @@
+const chainId = (network) => {
+  if (network === "rinkeby") {
+    return 4;
+  } else if (network === "mumbai") {
+    return 80001;
+  } else {
+    return 80001;
+  }
+};
+
+module.exports = chainId;
