+++ src/utils/txUrlStart.utils.js
@@ -1,0 +1,11 @@
+const { txUrl } = require("../data/network.data.js");
+
+const txUrlStart = (network) => {
+  if (network === "rinkeby") {
+    return txUrl.get("rinkeby");
+  } else if (network === "mumbai") {
+    return txUrl.get("mumbai");
+  }
+};
+
+module.exports = txUrlStart;
